07_pdf_rag_assistant/pyproject.toml
[project]
name = "07-pdf-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pypdf>=6.7.5",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
